requirements.txt
fastapi
uvicorn
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
streamlit
requests
langchain
langchain-community
langchain-core
langchain-openai
pydantic
groq
